feat: spacecat user-agent
