+++ src/sitemap/constants.js
@@ -1,0 +1,39 @@
+/*
+ * Copyright 2025 Adobe. All rights reserved.
+ * This file is licensed to you under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License. You may obtain a copy
+ * of the License at http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under
+ * the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
+ * OF ANY KIND, either express or implied. See the License for the specific language
+ * governing permissions and limitations under the License.
+ */
+
+// chrome mac browser agent
+export const DEFAULT_USER_AGENT = 'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/133.0.0.0 Safari/537.36';
+
+// spacecat user agent
+export const SPACECAT_USER_AGENT = 'spacecat/1.0';
+
+// sitemap error codes
+export const ERROR_CODES = Object.freeze({
+  INVALID_URL: 'INVALID URL',
+  NO_SITEMAP_IN_ROBOTS: 'NO SITEMAP FOUND IN ROBOTS',
+  NO_VALID_PATHS_EXTRACTED: 'NO VALID URLs FOUND IN SITEMAP',
+  SITEMAP_NOT_FOUND: 'NO SITEMAP FOUND',
+  SITEMAP_EMPTY: 'EMPTY SITEMAP',
+  SITEMAP_FORMAT: 'INVALID SITEMAP FORMAT',
+  FETCH_ERROR: 'ERROR FETCHING DATA',
+});
+
+// sitemap formats
+export const VALID_MIME_TYPES = Object.freeze([
+  'application/xml',
+  'text/xml',
+  'text/html',
+  'text/plain',
+]);
+
+// opportunities data
+export const TRACKED_STATUS_CODES = Object.freeze([301, 302, 404]);
