+++ src/sitemap/handler.js
@@ -28,28 +28,14 @@ import { AuditBuilder } from '../common/audit-builder.js';
 import { syncSuggestions } from '../utils/data-access.js';
 import { convertToOpportunity } from '../common/opportunity.js';
 import { createOpportunityData } from './opportunity-data-mapper.js';
+import {
+  DEFAULT_USER_AGENT,
+  ERROR_CODES,
+  VALID_MIME_TYPES,
+  TRACKED_STATUS_CODES,
+} from './constants.js';
 
 const auditType = Audit.AUDIT_TYPES.SITEMAP;
-// Add new constant for status codes we want to track
-const TRACKED_STATUS_CODES = Object.freeze([301, 302, 404]);
-const DEFAULT_USER_AGENT = 'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/131.0.0.0 Safari/537.36';
-
-export const ERROR_CODES = Object.freeze({
-  INVALID_URL: 'INVALID URL',
-  NO_SITEMAP_IN_ROBOTS: 'NO SITEMAP FOUND IN ROBOTS',
-  NO_VALID_PATHS_EXTRACTED: 'NO VALID URLs FOUND IN SITEMAP',
-  SITEMAP_NOT_FOUND: 'NO SITEMAP FOUND',
-  SITEMAP_EMPTY: 'EMPTY SITEMAP',
-  SITEMAP_FORMAT: 'INVALID SITEMAP FORMAT',
-  FETCH_ERROR: 'ERROR FETCHING DATA',
-});
-
-const VALID_MIME_TYPES = Object.freeze([
-  'application/xml',
-  'text/xml',
-  'text/html',
-  'text/plain',
-]);
 
 /**
  * Fetches the content from a given URL.
@@ -518,7 +504,6 @@ export function getSitemapsWithIssues(auditData) {
  * // Output:
  * // [
  * //   { type: 'url', sitemapUrl: 'https://site.url/sitemap.xml', pageUrl: 'https://site.url/page1', statusCode: 404 },
- * //   { type: 'url', sitemapUrl: 'https://site.url/sitemap.xml', pageUrl: 'https://site.url/page2', statusCode: 500 }
  * // ]
  */
 export function getPagesWithIssues(auditData) {
